баг в миграции

+++ database/migrations/2023_03_30_005300_create_orders_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->string("warehouseName");
             $table->string("oblast");
             $table->integer("incomeID");
-            $table->integer("odid");
+            $table->bigInteger("odid");
             $table->integer("nmId");
             $table->string("subject");
             $table->string("category");
